+++ src/main/java/puzzle6/RaceInfo.java
@@ -4,15 +4,15 @@ package puzzle6;
  * @param timeLimit milliseconds
  * @param record millimeters
  */
-public record RaceInfo(int timeLimit, int record) {
+public record RaceInfo(long timeLimit, long record) {
 
     //to hold or not to hold [the button]
-    public int numWaysToBeat() {
+    public long numWaysToBeat() {
 
         int winCount = 0;
-        for(int holdSeconds = 0; holdSeconds < timeLimit; holdSeconds++) {
-            int remainingTime = timeLimit - holdSeconds;
-            int distanceMoved = holdSeconds * remainingTime;
+        for(long holdSeconds = 0; holdSeconds < timeLimit; holdSeconds++) {
+            long remainingTime = timeLimit - holdSeconds;
+            long distanceMoved = holdSeconds * remainingTime;
 
             if(distanceMoved > record)
                 winCount++;
